Undo card image edits on cancel and reuse repeated panel code
