+++ front/app/containers/Admin/projects/_shared/components/ProjectSetupForm/ProjectContextSection/index.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 
 import { Box } from '@citizenlab/cl2-component-library';
 
-import useAuthUser from 'api/me/useAuthUser';
-
 import { SubSectionTitle } from 'components/admin/Section';
 
 import { FormattedMessage } from 'utils/cl-intl';
@@ -11,14 +9,12 @@ import { FormattedMessage } from 'utils/cl-intl';
 import Inner from './Inner';
 import messages from './messages';
 import { Props } from './types';
+import { useCanEditProjectContext } from './utils';
 
 const ProjectContextSection = (props: Props) => {
-  const { data: authUser } = useAuthUser();
+  const canEditProjectContext = useCanEditProjectContext();
 
-  if (!authUser) return null;
-  if (authUser.data.attributes.highest_role === 'project_moderator') {
-    return null;
-  }
+  if (!canEditProjectContext) return null;
 
   return (
     <Box mb="40px">
